backend/src/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & OCR
openai==1.3.5
pytesseract==0.3.10
pillow==10.1.0
opencv-python==4.8.1.78

# Data processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# File handling
python-magic==0.4.27
aiofiles==23.2.1

# Email & Notifications
fastapi-mail==1.4.1
celery==5.3.4
redis==5.0.1

# Payment processing
stripe==7.7.0

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.4.2
pydantic-settings==2.0.3
httpx==0.25.2
jinja2==3.1.2

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0